requirements.txt
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.2
Authlib==0.15.5
Babel==2.10.1
bcrypt==3.2.2
bidict==0.22.0
blinker==1.4
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
cargo==0.3
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==3.4.8
Django==4.0.4
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
eventlet==0.33.0
fastapi==0.78.0
Flask==2.1.2
Flask-Babel==2.0.0
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.4
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httpagentparser==1.9.2
httptools==0.4.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
ldap3==2.9.1
Mako==1.2.0
MarkupSafe==2.1.1
maturin==0.12.19
orjson==3.6.8
paramiko==2.11.0
passlib==1.7.4
path==16.4.0
pgadmin4==6.9
Pillow==9.1.1
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyNaCl==1.5.0
pyotp==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.2
python-jose==3.3.0
python-multipart==0.0.5
python-socketio==5.6.0
pytz==2021.3

PyYAML==6.0
qrcode==7.3.1
requests==2.25.1
rsa==4.8
s3transfer==0.5.2
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
speaklater3==1.4
SQLAlchemy==1.4.36
sqlparse==0.4.2
sshtunnel==0.4.0
starlette==0.19.1
to==0.3
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
ua-parser==0.10.0
ujson==5.3.0
urllib3==1.26.9
user-agents==2.2.0
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
Werkzeug==2.0.3
WTForms==3.0.1
